LayoutTests/imported/w3c/web-platform-tests/html/browsers/history/the-history-interface/history_forward_cross_realm_method-expected.txt



PASS history.forward() uses this's associated document's browsing context to determine if navigation is allowed
